server/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
PyMySQL==1.1.0
python-dotenv==1.0.0
flasgger==0.9.7.1
